requirements.txt
python-dotenv
transformers
diffusers
accelerate
protobuf
sentencepiece
tqdm
hf_xet
huggingface_hub